# docs/swagger.yaml
definitions:
...
  controllers.fundingListResBody:
    properties:
      achievementRate:
        type: number
      dDay:
        type: integer
      fundingName:
        type: string
      id:
        type: integer
      mainImgUrl:
        type: string
      sellerName:
        type: string
    type: object
  controllers.fundingResBody:
    properties:
      achievementRate:
        description: salesAmount / Price
        type: number
      artistName:
        type: string
      dDay:
        type: integer
      detailedImgUrl:
        type: string
      endDate:
        type: string
      fundName:
        type: string
      fundingImgUrls:
        items:
          type: string
        type: array
      id:
        type: integer
      price:
        type: integer
      salesAmount:
        type: integer
      sellerName:
        type: string
      startDate:
        type: string
      targetAmount:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /fundings:
    get:
      consumes:
      - application/json
      description: |-
        <br>아티스트와 관련된 펀딩 리스트를 반환합니다.
        <br>
      parameters:
      - description: .
        in: query
        name: artist-id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.fundingListResBody'
            type: array
      summary: 아티스트와 관련 펀딩 리스트
  /fundings/{fund_id}:
    get:
      consumes:
      - application/json
      description: |-
        <br>펀딩 상세정보를 반환합니다.
        <br>
        sellerName : 판매자의 닉네임<br>
        fundName : 펀드 이름<br>
        price : 하나를 구매할 때의 가격<br>
        targetAmount : 판매 목표량<br>
        salesAmount : 현재까지의 판매량<br>
        startDate : 펀딩 시작 일<br>
        endDate: 펀딩 마감 일<br>
        artistName : 펀딩과 관련된 연예인 이름<br>
        achievementRate : 펀딩 달성률 (판매량 / 목표량)<br>
        dDay : 펀딩 마감일까지 남은 날짜<br>
        fundingImgUrls : 펀딩 상품의 이미지 주소들<br>
        detailedImgUrl : 펀딩 상세정보 이미지<br>
      parameters:
      - description: fund_id
        in: path
        name: fund_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.fundingResBody'
      summary: 펀딩 상세정보
...
swagger: "2.0"